requirements.txt
mysql-connector-python==8.2.0
python-telegram-bot==22.3
